+++ FiyiStackDeskApp/Generators/G1/Modules/MSSQLServer/CreateTable.cs
@@ -19,7 +19,7 @@ SET QUOTED_IDENTIFIER ON
 --Last modification on: {DateTime.Now}
 
 CREATE TABLE [{TableScheme}].[{TableArea}.{TableName}] (
-    [{TableName}Id] [int] IDENTITY(1,1) NOT NULL,
+    [{TableName}Id] [bigint] IDENTITY(1,1) NOT NULL,
     CONSTRAINT [PK_{TableArea}{TableName}] PRIMARY KEY CLUSTERED ([{TableName}Id] ASC) WITH (PAD_INDEX = OFF, STATISTICS_NORECOMPUTE = OFF, IGNORE_DUP_KEY = OFF, ALLOW_ROW_LOCKS = ON, ALLOW_PAGE_LOCKS = ON)
 ON [PRIMARY])
 ON[PRIMARY]";
